Split v2 eval control helpers
